+++ inst/rmd/h/tocify/jquery.tocify.js
@@ -176,7 +176,7 @@
         self._setEventHandlers();
 
         // Binding to the Window load event to make sure the correct scrollTop is calculated
-        $(window).load(function() {
+        $(window).on("load", function() {
 
           // Sets the active TOC item
           self._setActiveElement(true);